pyproject.toml
[tool.poetry]
name = "fastapi-basic"
version = "0.1.0"
description = ""
authors = ["Cinnamoon-dev <pedrolcsilva10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
sqlalchemy = "^2.0.24"
psycopg2-binary = "^2.9.9"
httpx = "^0.26.0"
pytest = "^7.4.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
argon2-cffi = "^23.1.0"
fastapi-mail = "^1.4.1"
itsdangerous = "^2.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
